Completed HOC PostWrapper to make back buttons generic
