+++ src/article/postWrapper.css
@@ -1,0 +1,28 @@
+.post-wrapper .back-button {
+    display: block;
+    max-width: 800px;
+    width: 100%;
+    font-size: 1.2em;
+    font-family: "Roboto Slab", serif;
+    background-color: var(--secondary-background);
+    border-radius: 8px;
+    padding: 10px;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.post-wrapper .back-button > :first-child {
+    margin-right: 10px;
+}
+
+.post-wrapper .back-button:first-child {
+    margin-bottom: 10px;
+}
+
+.post-wrapper .back-button:last-child {
+    margin-top: 10px;
+}
+
+#bottom-back-button {
+    display: none;
+}
