+++ src/article/postWrapper.js
@@ -2,6 +2,20 @@ import React from 'react';
 import { Link } from "react-router-dom";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
+import "./postWrapper.css";
+
+function setBottomButtonVisibility(visible) {
+    const bottomButton = document.getElementById("bottom-back-button");
+    bottomButton.style.display = (visible) ? "block" : "none";
+}
+
+function updateBottomButtonVisibility() {
+    const wrapperContent = document.getElementsByClassName("post-wrapper")[0];
+    let wrapperContentHeight = wrapperContent.getBoundingClientRect().height;
+    let contentAreaHeight = wrapperContent.parentElement.parentElement.getBoundingClientRect().height;
+    setBottomButtonVisibility(wrapperContentHeight > contentAreaHeight);
+}
+
 /**
  * Takes a post (e.g. a blog post) and wraps it with back buttons above
  * and sometimes below
@@ -16,18 +30,37 @@ export default function postWrapper(WrappedComponent, url, name) {
             super(props);
         }
 
+        componentDidMount() {
+            updateBottomButtonVisibility();
+            // Shocking, another terrible idea
+            // Just adds the bottom button after the text should be loaded
+            // Need to replace this with promises/callbacks eventually
+            setTimeout(() => updateBottomButtonVisibility(), 550);
+            window.addEventListener("resize", () => {
+                updateBottomButtonVisibility();
+            });
+        }
+
         render() {
             return (
-                <>
+                <div class="post-wrapper">
                     <Link
-                        className={"blog-back-button"}
+                        className={"back-button"}
                         to={`${url}`}
                     >
                         <FontAwesomeIcon icon={"arrow-alt-circle-left"} />
-                        Back to { name }
+                        Back to {name}
                     </Link>
                     <WrappedComponent {...this.props} />
-                </>
+                    <Link
+                        className={"back-button"}
+                        id={"bottom-back-button"}
+                        to={`${url}`}
+                    >
+                        <FontAwesomeIcon icon={"arrow-alt-circle-left"} />
+                        Back to {name}
+                    </Link>
+                </div>
             );
         }
     }
