+++ src/blog.css
@@ -51,35 +51,6 @@
     margin-bottom: auto;
 }
 
-.blog-back-button {
-    display: block;
-    max-width: 800px;
-    width: 100%;
-    font-size: 1.2em;
-    font-family: "Roboto Slab", serif;
-    background-color: var(--secondary-background);
-    border-radius: 8px;
-    padding: 10px;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.blog-back-button > :first-child {
-    margin-right: 10px;
-}
-
-.blog-back-button:first-child {
-    margin-bottom: 10px;
-}
-
-.blog-back-button:last-child {
-    margin-top: 10px;
-}
-
-#blog-bottom-back-button {
-    display: none;
-}
-
 @media screen and (max-width: 500px) {
     .blog-link-date,
     .blog-link-icon {
